pyproject.toml
[project]
name = "unitares-discord-bridge"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
unitares-bridge = "bridge.bot:main"

[tool.pytest.ini_options]
# asyncio_mode=auto: pytest-asyncio treats every async test as an asyncio test
# automatically, preventing async tests from being silently skipped (issue #13)
asyncio_mode = "auto"
